[#235] Make possible to run DAL nodes
Problem: We always pass `--without-dal` in Kiln for the baker binary.
With the Rio protocol, using that flag or `--dal-node` became mandatory,
so it would be nice to allow the baker to run along a Data Availability
Layer node.

Solution: Attempt to parse a `--dal-node-uri` uri as an URI. If present,
we pass `--dal-node` to the baker binary with the provided URI,
otherwise we pass `--without-dal` as previously.