[#245] Fix snapshot import hanging
Problem: Importing a snapshot hangs for mainnet snapshot.

Solution: The problem appeared to be caused by a full stdout handle
which subsequently blocks the corresponding process. This was fixed
by duplicating the procedure that was currently run on stderr (the
backend expected the output of tezos node at stderr). As a precaution
the buffer contents are also printed to the log to ensure that it is
consumed, before attempting to parse it.